+++ backend/auth/current_user_context.py
@@ -1,5 +1,5 @@
 from dataclasses import dataclass
-from backend.repositories.schema.user import User
+from backend.schema.user import User
 
 @dataclass
 class CurrentUserContext:
