+++ backend/models/token.py
@@ -1,0 +1,11 @@
+from pydantic import BaseModel
+
+
+class Token(BaseModel):
+    access_token: str
+    token_type: str
+
+
+class TokenData(BaseModel):
+    user_id: str | None = None
+    permissions: set[str] | None = None
