--- backend/templates/email/password_reset.html
@@ -1,115 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-	<head>
-		<meta charset="utf-8">
-		<meta name="viewport" content="width=device-width, initial-scale=1.0">
-		<title>Reset Your Password</title>
-	</head>
-	<body style="margin: 0;
-				 padding: 0;
-				 font-family: Arial, Helvetica, sans-serif;
-				 background-color: #f4f4f4">
-		<table role="presentation"
-			   witdth="100%"
-			   ceilspacing="0"
-			   cellpadding="0"
-			   style="background-color: #f4f4f4;
-			   		  padding: 20px 0">
-			<tr>
-				<td align="center">
-					<table role="presentation"
-						   width="600"
-						   cellspacing="0"
-						   cellpadding="0"
-						   style="background-color: #ffffff;
-						   		  border-radius: 8px;
-								  overflow: hidden">
-						<!-- Header -->
-						<tr>
-							<td style="background-color: #527c9f; padding: 30px; text-align: center;">
-								<h1 style="margin: 0; color: #ffffff; font-size: 24px;">Florida Poly Campus Map</h1>
-							</td>
-						</tr>
-
-						<!-- Body -->
-						<tr>
-							<td style="padding: 40px 30px;">
-								<h2 style="margin: 0 0 20px 0; color: #333333; font-size: 20px;">Reset Your Password</h2>
-								<p style="margin: 0 0 15px 0;
-										  color: #555555;
-										  font-size: 16px;
-										  line-height: 1.5;">
-									Hi {{ username }},
-								</p>
-								<p style="margin: 0 0 25px 0;
-										  color: #555555;
-										  font-size: 16px;
-										  line-height: 1.5;">
-									You request to reset your password. Click the button below to set a new password:
-								</p>
-
-								<!-- Button -->
-								<table role="presentation"
-									   cellspacing="0"
-									   cellpadding="0"
-									   style="margin: 0 auto 25px auto">
-									<tr>
-										<td style="background-color: #527c9f; border-radius: 5px;">
-											<a href="{{ reset_url }}"
-											   style="display: inline-block;
-											   		  padding: 15px 30px;
-													  color: #ffffff;
-													  text-decoration: none;
-													  font-size: 16px;
-													  font-weight: bold">
-												Reset Password
-											</a>
-										</td>
-									</tr>
-								</table>
-								<p style="margin: 0 0 15px 0;
-										  color: #777777;
-										  font-size: 14px;
-										  line-height: 1.5">
-									This link will expire in <strong>{{ reset_token_expire_minutes }} minutes</strong>.
-								</p>
-								<p style="margin: 0 0 15px 0;
-										  color: #777777;
-										  font-size: 14px;
-										  line-height: 1.5">
-									If you didn't request this password reset, you can safely ignore this email. Your password will remain secure.
-								</p>
-
-								<!-- Fallback Link -->
-								<p style="margin: 25px 0 0 0;
-										  color: #999999;
-										  font-size: 12px;
-										  line-height: 1.5">
-									If the button doesn't work, copy and paste this link into your browser:
-									<br>
-									<a href="{{ reset_url }}" style="color: #527c9f; word-break: break-all;">{{ reset_url }}</a>
-								</p>
-							</td>
-						</tr>
-
-						<!-- Footer -->
-						<tr>
-							<td style="background-color: #f8f9fa;
-									   padding: 20px 30px;
-									   text-align: center;
-									   border-top: 1px solider #eeeeee">
-								<p style="margin: 0; color: #999999; font-size: 12px;">
-									This is an automated message from Florida Poly Campus Map.
-									<br>
-									Please do not reply to this email.
-									<br>
-									The Florida Poly Campus Map and its developers are not affliated with Florida Polytechnic University.
-								</p>
-							</td>
-						</tr>
-					</table>
-				</td>
-			</tr>
-		</table>
-	</body>
-</html>
